Run command as pseudo-entrypoint
The `ROCKER_ENTRYPOINT_COMMAND` specifies a `COMMAND` to run out of
the entrypoint script with command line args appended.

This is messy, because it's impossible to tell where `rocker` command
line args end and the `COMMAND` command line args begin, so it's
necessary to add the `--` arg as a signal.

This feature will become more useful when the `rocker` script is
called with different `$0`, and it can be assumed that all args will
pass to `COMMAND`.